+++ lib/page/settings_page/sync.dart
@@ -65,7 +65,7 @@ class _SyncSettingState extends ConsumerState<SyncSetting> {
                 child: GestureDetector(
                   onTap: () async {
                     if (!await launchUrl(
-                        Uri.parse('https://anx.anxcye.com/docs/sync/webdav'),
+                        Uri.parse('https://github.com/zym20192019/anx-reader-ai/blob/main/docs/sync.md'),
                         mode: LaunchMode.externalApplication)) {
                       AnxToast.show(L10n.of(context).commonFailed);
                     }